moved appbar to components
